Big changes:
Got it so that when ducks are pulled from MongoDB, the server will sort the ducks alphabetically automatically. This is because MongoDB doesn't sort them alphabetically on it's own, so when new ducks are added, they just show as the last on the list on the frontend. Not having them shown alphabetically pisses me off, so I had to fix that immediately.

I also did some filters for the other collections. Before, I didn't need to do it because I was working only on the "allducks" collection. Now that I'm branching into the others, I need to have those now. Right now, the only filters are by the type and brand, but others will be added s-needed. Do not add what you don't need yet.

I also changed the put request section so that information about ducks is only changed when you put something in the fields. If I didn't do this, then it would push out the blanks too, and then the server would complain about that. I also had to update the Joi so that it will accept the fields being optional when updating/editing only. If you're just adding ducks, then it's still required and will reject anything less.